#57 Moved test helpers
