+++ test_desktop/test_api_manager.cpp
@@ -2,7 +2,7 @@
 
 #include "../src/api/api_manager.h"
 #include "../src/test_dummys/api_manager_delegate_dummy.h"
-#include "../src/test_helpers/request_response_listener.h"
+#include "test_helpers/request_response_listener.h"
 #include "../src/api/protocol_paths.h"
 #include "../src/test_dummys/request_gadget_dummy.h"
 #include "../test_fixtures/api_jsons.h"
