+++ test_desktop/test_helpers/validator.cpp
